requirements.txt
streamlit
torch
torchvision
pandas
scikit-learn
pillow
git+https://github.com/openai/CLIP.git
